+++ sample/src/androidMain/kotlin/com/github/panpf/zoomimage/sample/AppSettings.android.kt
@@ -185,6 +185,10 @@ actual class AppSettings actual constructor(val context: PlatformContext) {
         booleanSettingsStateFlow(context, "delayImageLoadEnabled", false)
     }
 
+    actual val subsamplingEnabled: SettingsStateFlow<Boolean> by lazy {
+        booleanSettingsStateFlow(context, "subsamplingEnabled", true)
+    }
+
 
     // ------------------------------------------ other --------------------------------------------